requirements.txt
Django==1.10.0
djangorestframework==3.3.3
httpie